+++ form_bloc_web/lib/super_bloc_delegate.dart
@@ -1,6 +1,6 @@
 import 'package:bloc/bloc.dart';
 
-class SuperBlocDelegate extends BlocDelegate {
+class SuperBlocDelegate extends BlocObserver {
   @override
   void onEvent(Bloc bloc, Object event) {
     super.onEvent(bloc, event);
@@ -32,10 +32,10 @@ class SuperBlocDelegate extends BlocDelegate {
   }
 
   @override
-  void onError(Bloc bloc, Object error, StackTrace stacktrace) {
-    super.onError(bloc, error, stacktrace);
+  void onError(Cubit cubit, Object error, StackTrace stacktrace) {
+    super.onError(cubit, error, stacktrace);
     _printWrapped(
-      'bloc: ${bloc.runtimeType}, error: $error, stacktrace: $stacktrace',
+      'bloc: ${cubit.runtimeType}, error: $error, stacktrace: $stacktrace',
     );
   }