+++ packages/flutter_form_bloc/lib/src/dropdown_field_bloc_builder_mobile.dart
@@ -145,7 +145,7 @@ class _DropdownFieldBlocBuilderMobileState<Value>
       focusNode: _effectiveFocusNode,
       child: BlocBuilder<SelectFieldBloc<Value, dynamic>,
           SelectFieldBlocState<Value, dynamic>>(
-        bloc: widget.selectFieldBloc,
+        cubit: widget.selectFieldBloc,
         builder: (context, fieldState) {
           final isEnabled = fieldBlocIsEnabled(
             isEnabled: widget.isEnabled,
@@ -253,7 +253,7 @@ class _DropdownFieldBlocBuilderMobileState<Value>
   List<DropdownMenuItem<Value>> _buildItems(
     Iterable<Value> items,
   ) {
-    final style = Theme.of(context).textTheme.subhead.copyWith(
+    final style = Theme.of(context).textTheme.subtitle1.copyWith(
           color: ThemeData.estimateBrightnessForColor(
                       Theme.of(context).canvasColor) ==
                   Brightness.dark