+++ packages/flutter_form_bloc/lib/src/stepper/stepper_form_bloc_builder.dart
@@ -161,8 +161,8 @@ class StepperFormBlocBuilder<T extends FormBloc> extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return BlocBuilder<T, FormBlocState>(
-      bloc: formBloc,
-      condition: (p, c) =>
+      cubit: formBloc,
+      buildWhen: (p, c) =>
           p.numberOfSteps != c.numberOfSteps || p.currentStep != c.currentStep,
       builder: (context, state) {
         final formBloc = this.formBloc ?? context.bloc<T>();