+++ packages/flutter_form_bloc/lib/src/text_field_bloc_builder.dart
@@ -750,7 +750,7 @@ class _TextFieldBlocBuilderState extends State<TextFieldBlocBuilder> {
       animate: widget.animateWhenCanShow,
       builder: (_, __) {
         return BlocBuilder<TextFieldBloc, TextFieldBlocState>(
-          bloc: widget.textFieldBloc,
+          cubit: widget.textFieldBloc,
           builder: (context, state) {
             final isEnabled = fieldBlocIsEnabled(
               isEnabled: widget.isEnabled,
@@ -826,8 +826,8 @@ class _TextFieldBlocBuilderState extends State<TextFieldBlocBuilder> {
 
   Widget _buildTextField(
       {@required TextFieldBlocState state, @required bool isEnabled}) {
-    if (!kIsWeb) {
-      return TypeAheadField<String>(
+    return widgetBasedOnPlatform(
+      mobile: TypeAheadField<String>(
         textFieldConfiguration: TextFieldConfiguration<String>(
           controller: _controller,
           decoration: _buildDecoration(state),
@@ -843,7 +843,7 @@ class _TextFieldBlocBuilderState extends State<TextFieldBlocBuilder> {
                       .copyWith(color: Theme.of(context).disabledColor)
                   : Theme.of(context)
                       .textTheme
-                      .subhead
+                      .subtitle1
                       .copyWith(color: Theme.of(context).disabledColor),
           textAlign: widget.textAlign,
           textDirection: widget.textDirection,
@@ -910,7 +910,7 @@ class _TextFieldBlocBuilderState extends State<TextFieldBlocBuilder> {
                 child: Text(
                   'No Items Found!',
                   style: widget.suggestionTextStyle ??
-                      Theme.of(context).textTheme.subhead.copyWith(
+                      Theme.of(context).textTheme.subtitle1.copyWith(
                             color: ThemeData.estimateBrightnessForColor(
                                         Theme.of(context).canvasColor) ==
                                     Brightness.dark
@@ -937,7 +937,7 @@ class _TextFieldBlocBuilderState extends State<TextFieldBlocBuilder> {
             child: Text(
               suggestion,
               style: widget.suggestionTextStyle ??
-                  Theme.of(context).textTheme.subhead.copyWith(
+                  Theme.of(context).textTheme.subtitle1.copyWith(
                         color: ThemeData.estimateBrightnessForColor(
                                     Theme.of(context).canvasColor) ==
                                 Brightness.dark
@@ -965,9 +965,8 @@ class _TextFieldBlocBuilderState extends State<TextFieldBlocBuilder> {
             widget.textFieldBloc.selectSuggestion(suggestion);
           }
         },
-      );
-    } else {
-      return TextField(
+      ),
+      other: TextField(
         controller: _controller,
         decoration: _buildDecoration(state),
         keyboardType: widget.keyboardType,
@@ -981,7 +980,7 @@ class _TextFieldBlocBuilderState extends State<TextFieldBlocBuilder> {
                 ? widget.style.copyWith(color: Theme.of(context).disabledColor)
                 : Theme.of(context)
                     .textTheme
-                    .subhead
+                    .subtitle1
                     .copyWith(color: Theme.of(context).disabledColor),
         textAlign: widget.textAlign,
         textDirection: widget.textDirection,
@@ -1021,7 +1020,7 @@ class _TextFieldBlocBuilderState extends State<TextFieldBlocBuilder> {
         textAlignVertical: widget.textAlignVertical,
         toolbarOptions: widget.toolbarOptions,
         onTap: widget.onTap,
-      );
-    }
+      ),
+    );
   }
 }