+++ packages/form_bloc/lib/src/blocs/form_bloc_observer.dart
@@ -10,14 +10,14 @@ import 'form/form_bloc.dart';
 /// But it retains the behavior of the current [BlocSupervisor.delegate].
 ///
 /// You can customize using:
-/// * [FormBlocDelegate.notifyOnFieldBlocEvent].
-/// * [FormBlocDelegate.notifyOnFieldBlocTransition].
-/// * [FormBlocDelegate.notifyOnFieldBlocError].
-/// * [FormBlocDelegate.notifyOnFormBlocEvent].
-/// * [FormBlocDelegate.notifyOnFormBlocTransition].
-/// * [FormBlocDelegate.notifyOnFormBlocError].
-class FormBlocDelegate extends BlocDelegate {
-  final BlocDelegate _oldBlocDelegate;
+/// * [FormBlocObserver.notifyOnFieldBlocEvent].
+/// * [FormBlocObserver.notifyOnFieldBlocTransition].
+/// * [FormBlocObserver.notifyOnFieldBlocError].
+/// * [FormBlocObserver.notifyOnFormBlocEvent].
+/// * [FormBlocObserver.notifyOnFormBlocTransition].
+/// * [FormBlocObserver.notifyOnFormBlocError].
+class FormBlocObserver extends BlocObserver {
+  final BlocObserver _oldBlocObserver;
 
   /// If is `true` every `transition` of any
   /// [FieldBloc] will be notified
@@ -49,7 +49,7 @@ class FormBlocDelegate extends BlocDelegate {
   /// to [BlocSupervisor.delegate].
   static bool notifyOnFormBlocError = true;
 
-  FormBlocDelegate() : _oldBlocDelegate = BlocSupervisor.delegate;
+  FormBlocObserver() : _oldBlocObserver = Bloc.observer;
 
   @override
   void onEvent(Bloc bloc, Object event) {
@@ -65,8 +65,27 @@ class FormBlocDelegate extends BlocDelegate {
     }
 
     if (notify) {
-      _oldBlocDelegate.onEvent(bloc, event);
+      _oldBlocObserver.onEvent(bloc, event);
       super.onEvent(bloc, event);
+    }
+  }
+
+  @override
+  void onChange(Cubit cubit, Change change) {
+    var notify = true;
+
+    if ((cubit is SingleFieldBloc ||
+            cubit is GroupFieldBloc ||
+            cubit is ListFieldBloc) &&
+        !notifyOnFieldBlocTransition) {
+      notify = false;
+    } else if (cubit is FormBloc && !notifyOnFormBlocTransition) {
+      notify = false;
+    }
+
+    if (notify) {
+      _oldBlocObserver.onChange(cubit, change);
+      super.onChange(cubit, change);
     }
   }
 
@@ -84,46 +103,46 @@ class FormBlocDelegate extends BlocDelegate {
     }
 
     if (notify) {
-      _oldBlocDelegate.onTransition(bloc, transition);
+      _oldBlocObserver.onTransition(bloc, transition);
       super.onTransition(bloc, transition);
     }
   }
 
   @override
-  void onError(Bloc bloc, Object error, StackTrace stacktrace) {
+  void onError(Cubit cubit, Object error, StackTrace stacktrace) {
     var notify = true;
 
-    if ((bloc is SingleFieldBloc ||
-            bloc is GroupFieldBloc ||
-            bloc is ListFieldBloc) &&
+    if ((cubit is SingleFieldBloc ||
+            cubit is GroupFieldBloc ||
+            cubit is ListFieldBloc) &&
         !notifyOnFieldBlocError) {
       notify = false;
-    } else if (bloc is FormBloc && !notifyOnFormBlocError) {
+    } else if (cubit is FormBloc && !notifyOnFormBlocError) {
       notify = false;
     }
 
     if (notify) {
-      _oldBlocDelegate.onError(bloc, error, stacktrace);
-      super.onError(bloc, error, stacktrace);
+      _oldBlocObserver.onError(cubit, error, stacktrace);
+      super.onError(cubit, error, stacktrace);
     }
   }
 
   /// Override the current [BlocSupervisor.delegate]
-  /// with a [FormBlocDelegate] for hide `events` and `transitions`
+  /// with a [FormBlocObserver] for hide `events` and `transitions`
   /// of any [FieldBloc] or [FormBloc].
   ///
   /// But it retains the behavior of the current [BlocSupervisor.delegate].
   ///
   /// You can customize using:
-  /// * [FormBlocDelegate.notifyOnFieldBlocEvent].
-  /// * [FormBlocDelegate.notifyOnFieldBlocTransition].
-  /// * [FormBlocDelegate.notifyOnFieldBlocError].
-  /// * [FormBlocDelegate.notifyOnFormBlocEvent].
-  /// * [FormBlocDelegate.notifyOnFormBlocTransition].
-  /// * [FormBlocDelegate.notifyOnFormBlocError].
+  /// * [FormBlocObserver.notifyOnFieldBlocEvent].
+  /// * [FormBlocObserver.notifyOnFieldBlocTransition].
+  /// * [FormBlocObserver.notifyOnFieldBlocError].
+  /// * [FormBlocObserver.notifyOnFormBlocEvent].
+  /// * [FormBlocObserver.notifyOnFormBlocTransition].
+  /// * [FormBlocObserver.notifyOnFormBlocError].
   static void overrideDelegateOfBlocSupervisor() {
-    if (BlocSupervisor.delegate is! FormBlocDelegate) {
-      BlocSupervisor.delegate = FormBlocDelegate();
+    if (Bloc.observer is! FormBlocObserver) {
+      Bloc.observer = FormBlocObserver();
     }
   }
 }