+++ packages/form_bloc/test/field_bloc/field_bloc_test.dart
@@ -6,7 +6,7 @@ import 'package:test/test.dart';
 import '../utils/my_bloc_delegate.dart';
 
 void main() {
-  BlocSupervisor.delegate = MyBlocDelegate();
+  Bloc.observer = MyBlocObserver();
   // FormBlocDelegate.notifyOnFieldBlocEvent = true;
   // FormBlocDelegate.notifyOnFieldBlocTransition = true;
   group('FieldBloc:', () {
@@ -14,7 +14,6 @@ void main() {
       test('_initialValue is the value of the initial state.', () {
         InputFieldBloc<int, dynamic> fieldBloc;
         InputFieldBlocState<int, dynamic> initialState;
-        List<InputFieldBlocState> expectedStates;
 
         fieldBloc = InputFieldBloc<int, dynamic>(
           name: 'fieldName',
@@ -28,15 +27,9 @@ void main() {
           isValidating: false,
           name: 'fieldName',
         );
-        expectedStates = [initialState];
 
         expect(
-          fieldBloc,
-          emitsInOrder(expectedStates),
-        );
-
-        expect(
-          fieldBloc.initialState,
+          fieldBloc.state,
           initialState,
         );
 
@@ -52,15 +45,9 @@ void main() {
           isValidating: false,
           name: 'fieldName',
         );
-        expectedStates = [initialState];
 
         expect(
-          fieldBloc,
-          emitsInOrder(expectedStates),
-        );
-
-        expect(
-          fieldBloc.initialState,
+          fieldBloc.state,
           initialState,
         );
       });
@@ -104,7 +91,7 @@ void main() {
         );
 
         final expectedStates = [
-          state1,
+          // state1,
           state2,
           state3,
           state4,
@@ -159,7 +146,7 @@ void main() {
         );
 
         final expectedStates = [
-          state1,
+          // state1,
           state2,
           state3,
         ];
@@ -190,8 +177,8 @@ void main() {
         ];
 
         expect(
-          fieldBloc,
-          emitsInOrder(expectedStates),
+          fieldBloc.state,
+          expectedStates.first,
         );
 
         expect(
@@ -226,18 +213,9 @@ void main() {
         name: 'fieldName',
       );
 
-      final expectedStates = [
-        initialState,
-      ];
-
       expect(
-        fieldBloc.initialState,
+        fieldBloc.state,
         initialState,
-      );
-
-      expect(
-        fieldBloc,
-        emitsInOrder(expectedStates),
       );
     });
 
@@ -271,7 +249,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
         state4,
@@ -313,7 +291,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
         state4,
@@ -365,7 +343,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
         state4,
@@ -406,7 +384,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
       ];
 
@@ -463,7 +441,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
       ];
@@ -510,7 +488,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
         state4,
@@ -551,7 +529,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
       ];
@@ -598,7 +576,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
       ];
@@ -655,7 +633,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
         state4,
@@ -725,7 +703,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
         state4,
@@ -781,7 +759,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
       ];
@@ -841,7 +819,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
         state4,
@@ -911,7 +889,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
         state4,
@@ -953,7 +931,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
       ];
       expect(
@@ -983,7 +961,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
       ];
       expect(
@@ -1021,7 +999,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
       ];
       expect(
@@ -1083,7 +1061,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
         state4,
@@ -1146,7 +1124,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
         state4,
@@ -1156,7 +1134,7 @@ void main() {
         emitsInOrder(expectedStates),
       );
 
-      fieldBloc.addError('error');
+      fieldBloc.addFieldError('error');
       await Future<void>.delayed(Duration(milliseconds: 0));
       fieldBloc.updateValue(1);
       fieldBloc.updateValue(null);
@@ -1194,7 +1172,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state4,
         state5,
@@ -1204,7 +1182,7 @@ void main() {
         emitsInOrder(expectedStates),
       );
 
-      fieldBloc.addError('error', isPermanent: true);
+      fieldBloc.addFieldError('error', isPermanent: true);
       fieldBloc.updateValue(2);
       fieldBloc.updateValue(1);
     });
@@ -1244,7 +1222,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
         state4,
@@ -1260,7 +1238,7 @@ void main() {
         autoValidate: false,
       ));
 
-      fieldBloc.addError('error', isPermanent: true);
+      fieldBloc.addFieldError('error', isPermanent: true);
       fieldBloc.updateValue(2);
       fieldBloc.updateValue(1);
     });
@@ -1298,7 +1276,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
         state4,
@@ -1314,7 +1292,7 @@ void main() {
         autoValidate: false,
       ));
 
-      fieldBloc.addError('error');
+      fieldBloc.addFieldError('error');
       await Future<void>.delayed(Duration(milliseconds: 0));
       fieldBloc.updateValue(1);
       fieldBloc.updateValue(null);
@@ -1372,7 +1350,7 @@ void main() {
     });
 
     test('updateExtraData method', () async {
-      final expected = [null, 1];
+      final expected = [1];
 
       final fieldBloc = InputFieldBloc<int, int>();