+++ packages/form_bloc/test/select_field/select_field_bloc_test.dart
@@ -36,7 +36,7 @@ void main() {
         );
 
         final expectedStates = [
-          state1,
+          // state1,
           state2,
         ];
         expect(
@@ -51,7 +51,6 @@ void main() {
     test('initial state.', () {
       SelectFieldBloc fieldBloc;
       SelectFieldBlocState initialState;
-      List<SelectFieldBlocState> expectedStates;
 
       fieldBloc = SelectFieldBloc<bool, dynamic>(
         name: 'name',
@@ -68,16 +67,9 @@ void main() {
         items: [],
       );
 
-      expectedStates = [initialState];
-
       expect(
-        fieldBloc.initialState,
+        fieldBloc.state,
         initialState,
-      );
-
-      expect(
-        fieldBloc,
-        emitsInOrder(expectedStates),
       );
 
       fieldBloc.close();
@@ -100,16 +92,9 @@ void main() {
         items: [true, false],
       );
 
-      expectedStates = [initialState];
-
       expect(
-        fieldBloc.initialState,
+        fieldBloc.state,
         initialState,
-      );
-
-      expect(
-        fieldBloc,
-        emitsInOrder(expectedStates),
       );
     });
 
@@ -136,7 +121,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
       ];
@@ -156,13 +141,13 @@ void main() {
         items: [true, false],
       );
 
-      final expectedState = fieldBloc.initialState.copyWith(
+      final expectedState = fieldBloc.state.copyWith(
         value: Optional.absent(),
         items: Optional.of([false]),
       );
 
       expect(
-        fieldBloc.skip(1),
+        fieldBloc,
         emitsInOrder(<SelectFieldBlocState>[expectedState]),
       );
 
@@ -196,7 +181,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
         state4,
@@ -235,7 +220,7 @@ void main() {
       );
 
       final expectedStates = [
-        state1,
+        // state1,
         state2,
         state3,
       ];
@@ -256,13 +241,13 @@ void main() {
         items: [true, false],
       );
 
-      final expectedState = fieldBloc.initialState.copyWith(
+      final expectedState = fieldBloc.state.copyWith(
         value: Optional.absent(),
         items: Optional.of([false]),
       );
 
       expect(
-        fieldBloc.skip(1),
+        fieldBloc,
         emitsInOrder(<SelectFieldBlocState>[expectedState]),
       );