+++ packages/flutter_form_bloc/CHANGELOG.md
@@ -1,3 +1,25 @@
+# 0.19.0
+  ## Breaking changes
+  * Updated to `form_bloc: ^0.18.0`.
+  * Updated to `flutter_bloc: ^6.0.0`.
+
+# 0.18.0
+Bad version, you should use `>=0.19.0` or `<= 0.15.1`.
+
+# 0.17.0
+Bad version, you should use `>=0.19.0` or `<= 0.15.1`.
+
+# 0.16.0
+Bad version, you should use `>=0.19.0` or `<= 0.15.1`.
+
+# 0.15.1
+  * Fix widgets in web platform.
+
+# 0.15.0
+  ## Breaking changes
+  * Updated to `form_bloc: ^0.15.0`.
+  * Updated to `flutter_bloc: ^0.15.1`.
+
 # 0.14.0
 * Updated to `form_bloc: ^0.14.0`.
 * Updated to `flutter_bloc: ^0.4.0`.